+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 2.2.0 - 2017-10-23
 
+- Fix Fiji GeoJSON extent using optimal tile cover
 - Support GeoJSON maximum/optimal tile cover
 
 ## 2.1.0 - 2017-03-30